print gcc errors in run.sh if compilation fails

+++ run.sh
@@ -1,6 +1,14 @@
 #!/usr/bin/env bash
 
 # recompile to update rom (in rom.h)
-gcc scripts/main.c scripts/emulate.c scripts/simulate.c scripts/write_rom.c scripts/util.c -lngspice > /dev/null 2>&1
-./a.out "$@"
-rm a.out
+gcc_output=$(gcc scripts/main.c scripts/emulate.c scripts/simulate.c scripts/write_rom.c scripts/util.c -lngspice 2>&1)
+if [ $? -eq 0 ]
+then
+    # compilation succeeded
+    ./a.out "$@"
+    rm a.out
+else
+    # compilation failed, print errors
+    echo $gcc_output
+    exit
+fi
